fix: move util
