+++ src/cli/typescript/typescript.ts
@@ -1,8 +1,6 @@
 import execa, { ExecaError } from 'execa';
-import path from 'path';
-import yargs from 'yargs';
-import { hideBin } from 'yargs/helpers';
 import { ConfigInternal } from '../../common/types';
+import { getArgs } from '../../common/utils';
 
 /**
  * Retrieves and displays the resolved TypeScript configuration from TSC.
@@ -65,45 +63,6 @@ export const compile = async (): Promise<string> => {
   }
 
   return compilerOutputCache;
-};
-
-/**
- * Parses command-line arguments to extract TypeScript configuration details.
- *
- * This function processes the command-line arguments using `yargs`, identifies any custom
- * TypeScript configuration file specified via the `tsconfig` argument, and formats
- * other arguments for use with the TypeScript Compiler (TSC). Additionally, it resolves
- * the full path to the specified or default `tsconfig.json` file.
- *
- * @returns {Object} An object containing:
- *   - `argv` {string[]}: An array of arguments formatted for TSC (e.g., `--key value`),
- *     excluding certain defaults like `_`, `$0`, `project`, and `tsconfig`.
- *   - `tsConfigFile` {string}: The name of the TypeScript configuration file to use,
- *     defaulting to `tsconfig.json` if no custom file is specified.
- *   - `tsConfig` {string}: The absolute path to the resolved TypeScript configuration file.
- */
-const getArgs = (): { argv: string[]; tsConfigFile: string; tsConfig: string } => {
-  let tsConfigFile: string = 'tsconfig.json';
-  const args = yargs(hideBin(process.argv)).parse();
-  const argv = Object.entries(args).reduce<string[]>((acc, arg) => {
-    const [key, value] = arg;
-
-    if (key === 'tsconfig') {
-      tsConfigFile = value;
-    }
-
-    if (key !== '_' && key !== '$0' && key !== 'project' && key !== 'tsconfig') {
-      acc.push(...[`--${key}`, value]);
-    }
-
-    return acc;
-  }, []);
-
-  return {
-    argv,
-    tsConfigFile,
-    tsConfig: `${path.resolve()}/${tsConfigFile}`,
-  };
 };
 
 function isExecaError(error: unknown): error is ExecaError {
